Fix ORM relationship errors, add compose.yml, ensure schema-ORM alignment
ORM fixes:
- models.py: Fix SQLAlchemy reserved word — rename Event.metadata and Classification.metadata Python attrs to extra_data (SQL column name unchanged)
- models.py: Player.player_aliases — add primaryjoin for non-FK steam_id relationship
- models.py: Team.match_teams — add foreign_keys to disambiguate team_id vs parent_team_id FKs
- models.py: Kill↔Round — no FK exists; document with explicit string primaryjoin using foreign() annotation; use viewonly=True
- models.py: Add foreign and and_ imports; Classification.entity_id → BigInteger (64-bit Steam IDs)
- models.py: Add alembic env.py schema creation before alembic_version table creation

Migration fixes:
- 0001_initial_schema.py: Remove duplicate played_at columns (already in CREATE TABLE)
- 0001_initial_schema.py: Remove redundant CREATE SCHEMA (now handled in env.py)
- 0002: entity_id → BigInteger to match ORM

Alembic env.py:
- Add schema creation before alembic_version table to support fresh DB installs
- Import text from sqlalchemy for proper SQL execution

compose.yml:
- New Docker Compose with postgres:16-alpine + adminer (tools profile)
- User/pass: csda/csda, DB: csda, port 5432

.env:
- DATABASE_URL=postgresql://csda:csda@localhost:5432/csda
- .env.example updated with same default

DB verified working:
- Fresh docker compose up postgres + alembic upgrade head ✓
- Full round-trip: Event → EventSeries → Match → Round → Kill → ClassifierRun (event+match scope) → Classification ✓
- 64-bit Steam IDs (BigInteger entity_id) stored and retrieved correctly ✓
- 29 tables all present ✓

+++ alembic/versions/0002_events_series_classifications.py
@@ -101,7 +101,7 @@ def upgrade() -> None:
         sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
         sa.Column("classifier_run_id", sa.Integer(), sa.ForeignKey("csda.classifier_runs.id"), nullable=False),
         sa.Column("entity_type", sa.Text(), nullable=False),
-        sa.Column("entity_id", sa.Integer(), nullable=False),
+        sa.Column("entity_id", sa.BigInteger(), nullable=False),
         sa.Column("label_name", sa.Text(), nullable=False),
         sa.Column("label_value", sa.Text(), nullable=False),
         sa.Column("confidence", sa.Numeric(4, 3), nullable=True),
